Rename to LivePublicationService
... bc it hopefully makes things clearer.
