+++ modules/live-publication-impl/src/test/resources/assetmanager-mp-live.xml
