+++ modules/live-publication-impl/src/test/resources/search-mp-live.xml
