+++ modules/scheduler-impl/pom.xml
@@ -186,7 +186,7 @@
     </dependency>
       <dependency>
           <groupId>org.opencastproject</groupId>
-          <artifactId>opencast-live-schedule-api</artifactId>
+          <artifactId>opencast-live-publication-api</artifactId>
           <version>19-SNAPSHOT</version>
           <scope>compile</scope>
       </dependency>
